Extract fibbonacci cairo1 test logic to a separate fn
