+++ src/utils.rs
@@ -522,6 +522,116 @@ pub mod test_utils {
     pub(crate) use run_syscall_hint;
 }
 
+pub mod integration_test_utils {
+    use crate::{
+        business_logic::{
+            execution::{
+                execution_entry_point::ExecutionEntryPoint,
+                objects::{CallInfo, CallType, TransactionExecutionContext},
+            },
+            fact_state::{
+                in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
+            },
+            state::cached_state::CachedState,
+        },
+        definitions::{constants::TRANSACTION_VERSION, general_config::StarknetGeneralConfig},
+        services::api::contract_classes::deprecated_contract_class::EntryPointType,
+        utils::ClassHash,
+    };
+    use num_traits::Zero;
+    use std::collections::HashMap;
+
+    use crate::utils::Address;
+    use cairo_lang_starknet::casm_contract_class::CasmContractClass;
+    use cairo_vm::{felt::Felt252, vm::runners::cairo_runner::ExecutionResources};
+
+    pub fn test_cairo_1_external_entrypoint(
+        program_data: &[u8],
+        entrypoint: usize,
+        expected_retdata: &[Felt252],
+        expected_execution_resources: &ExecutionResources,
+    ) {
+        let contract_class: CasmContractClass = serde_json::from_slice(program_data).unwrap();
+        let entrypoints = contract_class.clone().entry_points_by_type;
+        let entrypoint_selector = &entrypoints.external.get(entrypoint).unwrap().selector;
+
+        // Create state reader with class hash data
+        let mut contract_class_cache = HashMap::new();
+
+        let address = Address(1111.into());
+        let class_hash: ClassHash = [1; 32];
+        let nonce = Felt252::zero();
+
+        contract_class_cache.insert(class_hash, contract_class);
+        let mut state_reader = InMemoryStateReader::default();
+        state_reader
+            .address_to_class_hash_mut()
+            .insert(address.clone(), class_hash);
+        state_reader
+            .address_to_nonce_mut()
+            .insert(address.clone(), nonce);
+
+        // Create state from the state_reader and contract cache.
+        let mut state = CachedState::new(state_reader, None, Some(contract_class_cache));
+
+        // Create an execution entry point
+        let calldata = [0.into(), 1.into(), 12.into()].to_vec();
+        let caller_address = Address(0000.into());
+        let entry_point_type = EntryPointType::External;
+
+        let exec_entry_point = ExecutionEntryPoint::new(
+            address,
+            calldata.clone(),
+            Felt252::new(entrypoint_selector.clone()),
+            caller_address,
+            entry_point_type,
+            Some(CallType::Delegate),
+            Some(class_hash),
+            100000,
+        );
+
+        // Execute the entrypoint
+        let general_config = StarknetGeneralConfig::default();
+        let tx_execution_context = TransactionExecutionContext::new(
+            Address(0.into()),
+            Felt252::zero(),
+            Vec::new(),
+            0,
+            10.into(),
+            general_config.invoke_tx_max_n_steps(),
+            TRANSACTION_VERSION,
+        );
+        let mut resources_manager = ExecutionResourcesManager::default();
+
+        // expected results
+        let expected_call_info = CallInfo {
+            caller_address: Address(0.into()),
+            call_type: Some(CallType::Delegate),
+            contract_address: Address(1111.into()),
+            entry_point_selector: Some(Felt252::new(entrypoint_selector)),
+            entry_point_type: Some(EntryPointType::External),
+            calldata,
+            retdata: expected_retdata.to_vec(),
+            execution_resources: expected_execution_resources.clone(),
+            class_hash: Some(class_hash),
+            ..Default::default()
+        };
+
+        assert_eq!(
+            exec_entry_point
+                .execute(
+                    &mut state,
+                    &general_config,
+                    &mut resources_manager,
+                    &tx_execution_context,
+                    false,
+                )
+                .unwrap(),
+            expected_call_info
+        );
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
